+++ src/components/SearchHeader/index.js
@@ -1,58 +1,77 @@
 import Taro, { Component } from '@tarojs/taro'
 import { View, Picker, Input, Button } from '@tarojs/components'
-import { AtIcon, AtButton } from 'taro-ui'
+import { AtIcon } from 'taro-ui'
 import PropTypes from 'prop-types'
 
 import './index.scss'
 
 class SearchHeader extends Component {
   static propTypes = {
-    dateStart: PropTypes.string,
-    dateEnd: PropTypes.string,
-    model: PropTypes.object,
-    onDateChange: PropTypes.func,
-    onBtnDateClick: PropTypes.func
+    searchTypes: PropTypes.array,
+    onHandleSearch: PropTypes.func
+  }
+
+  constructor() {
+    super(...arguments)
+
+    this.state = {
+      searchKeyword: '',
+      selectedType: {
+        key: '搜索A',
+        value: 'searchA'
+      }
+    }
   }
 
   static defaultProps = {
     searchTypes: [
       {
-        key: '客户名',
-        value: 'customerName'
+        key: '搜索A',
+        value: 'searchA'
       },
       {
-        key: '电话',
-        value: 'customerPhone'
+        key: '搜索B',
+        value: 'searchB'
       },
       {
-        key: '备注',
-        value: 'customerRemark'
+        key: '搜索C',
+        value: 'searchB'
       }
-    ],
-    searchType: {
-      key: '客户名',
-      value: 'customerName'
+    ]
+  }
+
+  componentDidMount() {
+    const { searchTypes } = this.props
+    if (Array.isArray(searchTypes) && searchTypes.length > 0) {
+      this.setState({ selectedType: searchTypes[0] })
     }
   }
 
-  handleBtnDateClick = (btnName) => {
-    this.props.onBtnDateClick(btnName)
+  handleInput = ({ detail }) => {
+    const { value: searchKeyword } = detail
+    this.setState({ searchKeyword })
+  }
+
+  handlePickerChange = ({ detail }) => {
+    const { value: index } = detail
+    this.setState({ selectedType: this.props.searchTypes[index] })
   }
 
   render() {
-    const { searchTypes, searchType, keyword } = this.props
+    const { selectedType, searchKeyword } = this.state
+    const { searchTypes, onHandleSearch } = this.props
     return (
       <View className='header'>
         <View className='search-content'>
-          <Picker style='padding-right:10px;'  className='search-type' mode='selector' range={searchTypes} rangeKey='key' onChange={this.onTypeChange}>
+          <Picker style='padding-right:4px;' className='search-type' mode='selector' range={searchTypes} rangeKey='key' onChange={this.handlePickerChange}>
             <View className='customer-type'>
-              {searchType.key}
+              {selectedType.key}
               <AtIcon value='chevron-down' size='24' color='rgba(117, 117, 119, 1)'></AtIcon>
             </View>
           </Picker>
-          <Input className='input-content' type='string' name='keyword' placeholder='请输入查询关键字' value={keyword} onInput={this.handleKeyword} />
+          <Input onInput={this.handleInput} className='input-content' type='string' name='keyword' placeholder='请输入查询关键字' value={searchKeyword} />
         </View>
-        <Button className='btn-submit' type='primary' size='small'>搜索</Button>
+        <Button className='btn-submit' type='primary' size='small' onClick={onHandleSearch.bind(this, selectedType.value, searchKeyword)}>搜索</Button>
       </View>
     )
   }